requirements.txt
tensorflow==2.8.0
pandas==1.4.1
yfinance==0.1.70
matplotlib==3.5.1
numpy==1.22.0
seaborn==0.11.2
PyYAML==6.0